Removing non tutorial examples from DataAtWork
I'm curious to know what people think about removing these usage examples. Our spec says that DataAtWork is for "A list of links to examples of the dataset being used. May include tutorials, demos, or applications." The papers posted here fit the definition of examples of the data being used, but they're behind a paywall. 

Perhaps we should update the spec to be clearer about what we'd like to see here. "A list of links to examples of the dataset being used on AWS. May include tutorials, demos, or applications."