Add a new allen-ivy-glioblastoma-atlas dataset entry
This dataset consists of images of glioblastoma human brain tumor
tissue sections that have been probed for expression of particular genes
believed to play a role in development of the cancer.  Each tissue
section is adjacent to another section that was stained with a reagent
useful for identifying histological features of the tumor. Each of
these types of images has been completely annotated for tumor features
by a machine learning process trained by expert medical doctors.