Update and add Allen Institute for Brain Science datasets (#1035)
* Update allen-mouse-brain-atlas dataset entry

The https://github.com/AllenInstitute/open_dataset_tools
repository has recently been refactored and the location of the example
IPython notebook has been changed.

This commit updates the yaml entry to correctly point to the new
updated location of the example notebook.

* Add a new allen-ivy-glioblastoma-atlas dataset entry

This dataset consists of images of glioblastoma human brain tumor
tissue sections that have been probed for expression of particular genes
believed to play a role in development of the cancer.  Each tissue
section is adjacent to another section that was stained with a reagent
useful for identifying histological features of the tumor. Each of
these types of images has been completely annotated for tumor features
by a machine learning process trained by expert medical doctors.

* Add "glioblastoma" to the list of supported dataset tags