Removing the rechunked version of this dataset
An issue was detected with the data in this (rechunked) version of the zarr store which implicates the data is incorrect.

The tests are in this notebook: https://github.com/abarciauskas-bgse/2022-esip-kerchunk-tutorial/blob/main/MURSST-03-Tests.ipynb where the mean of the data for one day and a 5 degree x 5 degree cell is calculated for a kerchunked version of the zarr store and compared with each version of the zarr store in the AWS public datasets bucket. This calculation indicates that the data in the /zarr directory is incorrect. Upon inspection of the data, it seems like some of the data _should_ be masked as `NaN` but in the rechunked version of the data store (the data in `/zarr` with 6443 time steps per chunk) the data is not masked (i.e. there are values where we would expect `NaN`s)