Add YAML file for SISSA Forecast database
This YAML file contains data and description of the recently retrospective daily forecast made in SISSA. As some data is still being calculated, the data will be uploaded when its finish. Also the website with the final form of tutorial, will be uploaded soon, so for now, there is only a mail to contact us in case of questions.